requirements.txt
Cython==0.25.2
Django==1.10.6
Pygments==2.2.0
amqp==2.1.4
appdirs==1.4.0
billiard==3.5.0.2
celery==4.0.2
crcmod==1.7
decorator==4.0.11
django-angular==0.8.4
django-channels==0.7.0
django-cors-headers==2.0.2
django-crispy-forms==1.6.1
django-debug-toolbar==1.6
django-extensions==1.7.7
django-filter==1.0.1
django-filters==0.2.1
django-jet==1.0.5
django-nose==1.4.4
django-redis-sessions==0.5.6
djangorestframework==3.5.4
ipython==5.2.1
ipython-genutils==0.1.0
kombu==4.0.2
nanomsg==1.0
netaddr==0.7.19
nose==1.3.7
numexpr==2.6.2
numpy==1.12.0
oauthlib==2.0.1
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.10
protobuf==3.2.0
psycopg2==2.7
ptyprocess==0.5.1
pyparsing==2.1.10
pytz==2016.10
pyzmq==16.0.2
redis==2.10.5
requests==2.13.0
requests-oauthlib==0.8.0
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.2.3
tables==3.3.0
traitlets==4.3.1
vine==1.1.3
wcwidth==0.1.7
